The remaineder of the last commit message with a meaningless change cause I dun-goofed and copmmited and pushed before finishing the commit message.
Replaced UserInput class with something similar to the UserMultiInput class and removed the UserMultiInputClass. The UserInput class now has an indexer by which individual properties of the TargetValue element can be bound to and edited. There is still a single value converter to the class (something that may be changed in future) but properties of the TargetValue can be accessed individually now.

Made changes to the GameEditingData class to implement the new version of the UserInput class.Changed UserInputValidatorError's Properties to have mostly private init setters (instead of pulbic init) and as such require necessary values to be passed to the (also newly added) constructor.