Formatting changes to SimpleStringInput.cs.
Added default value to identifier parameter of StrignInputValidationError constructor.

Removed the asdf project that was involved in the fixing of a type in the name of the QuickTestingApplication project.

Added MetaTabView into the MetaTab of the TabControl in MainWindow.

Set the DataContext of MetaTabView to a new GameEditingData instance in the constructor of Meta TabView.

Added TestInt of type SimpleInputString property to GameEditingData (and a corresponding ValueConverter) added a TextBox to MetaTabView bound to that property of GameEditingData.