Initialized SimpleStringInput._ErrorsList at the declaration. Added nullable ? markers to SimpleStringInput.
Added call PropertyChanged on ErrorLevel in SimpleStringInput.OnErrorsListChanged.

Changed ISimpleStringInput.ErrorLevel to be abstract because INotifyPropertyChanged does not work with members inherited from interfaces. Added ErrorLevel property to SimpleStringInput<T>.

Added None error level to StringInputValidationErrorSeverity enum.

Created ErrorSeverityGreaterThanConverter IValueConverter to determine if an ErrorSeverity is greater than the conversion prameter (also an ErrorSeverity).

Attempted to debug ErrorSeverityToColorConverter. It is returning the right value but it's just not showing it in the app.