Renamed StringInputValidationError.cs and the corresponding class to ValidationError.cs. Changed ValidationError from using the enum StringInputValidationErrorSeverity to being a generic type and taking an enum type as a type argument to use as the error severity enum.
Deleted the WPFUtilities.StringInputValidationErrorSeverity enum and replaced it with ErrorSeverity (in a new file) in CCSSDomain.

Renamed ErrorSeverityGreaterThanConverter.cs and the corresponding class to EnumGreaterThanConverter.cs. Made EnumGreaterThanConverter a generic taking an Enum type as a type argument and changed the implementation of the converters so that they can compare values of that enum type.

Created ErrorSeverityGreaterThanConverter.cs and a corresponding class that implements EnumGreaterThanConverter<CCSSDomain.ErrorSeverity>.

Created IErrorConverter.cs and a corresponding interface in WPFUtilities.

Changed SimpleStringInput so that instead of using StringInputValidationError it is a generic and takes an enum as a type arguement. Changed the SimpleStringInputConverter to take an enum as another type parameter to use as the type arguement for ValdiationError<TErrorSeverity>.

Commented out the contents of StringInput.cs and commented out members of GameEditingData that used it.

Updated members of GameEditingData to comply with the changes to ValidationError, ErrorSeverity, and SimpleInputString.

Changed ErrorToBrushConverter to match the updates in IErrorConveter<>.