Renamed SimpleStringInput and related types and interfaces to StringInput.
Removed the usage of IStringInput because everything works without it and I forget why I needed it. Commented out the definition of IStrignInput.

Fixed possible null reference exception in StringInput.ErrorLevel.

Added private setter to StringInput.TargetObject to be used in StringInput.ConvertValue.

Renamed StringInput.ConvertValue to ConvertInputString and made it use InputString instead of taking an inputString parameter.

Added a comment to IErrorConverter.