Added SingleInputInverter and 16 generic overloads of MultiInputInverter to InputDelegates.cs.
Changed OutputObject to have a public setter in IInput<> and Input<,> in InputAbstractions.cs.

Some property name changes and access modifier changes in MutliInput.cs. Also added 4 generic overloads of the updated MultiInput.

Modified the constructor overloads of SingleInput. Changed some names in SingleInput. Changed the setter of SingleInput.OutputObject to be public.

Updated GameEditingData and AllianceEditingData and the corresponding validators to work with the updated inputs.