Added a non generic class Option with a NoValue field and added a implicit conversion in Optional<> between the Optional.NoValue and Option<>.NoVale. Changed Optional<> from a struct to a class for consistency.
Changed the converters, inverters, and validators from using nullable return types to using optional return types. Did the same for MutliInput, SingleInput, and ConversionPair.

Changed the methods AllianceValidator and GameValidator to use the optional return types. Also updated some of the GameValidator methods to use a single error return instead of an error list.

Changed ValidationData, AllianceValidator, and GameValidator to use a type alias of "Error" for ValidationError<ErrorSeverity>.